# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.5"
  title: SiB Colombia - Dataportal API
  description: API for SiB Colombia Dataportal. This API is used as a backend for SiB Colombia's web sites and online services.
  termsOfService: http://swagger.io/terms/
  contact:
    name: SiB Colombia Developer Team
    url: http://www.sibcolombia.net
    email: lgrajales@humboldt.org.co
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
# during dev, should point to your local machine
host: localhost:5000
...
basePath: /api/v1.5
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /registry/department:
    # binds a127 app logic to a route
    x-swagger-router-controller: registries
    get:
      description: Returns dataportal departments list
      tags:
        - registry
      # used as the method name of the controller
      operationId: departmentList
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/department'
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /registry/county:
    # binds a127 app logic to a route
    x-swagger-router-controller: registries
    get:
      description: Returns dataportal counties list
      tags:
        - registry
      # used as the method name of the controller
      operationId: countyList
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/county'
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /provider:
    # binds a127 app logic to a route
    x-swagger-router-controller: occurrences
    get:
      description: Returns dataportal provider list
      tags:
        - occurrence
      # used as the method name of the controller
      operationId: providerList
      parameters:
        - name: providerId
          in: query
          description: Provider Id.
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OccurrenceSearchResults"
        # responses may fall through to errors
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  OccurrenceSearchResults:
    required:
      - offset
      - size
      - count
      - results
    properties:
      offset:
        type: integer
      size:
        type: integer
      count:
        type: integer
      results:
        type: array
        items:
          $ref: '#/definitions/Occurrence'
      facets:
        type: array
        items:
          $ref: '#/definitions/Facet'
  Occurrence:
    required:
      - id
      - dwca_id
      - sourcefileid
      - occurrenceid
      - canonical
    properties:
      id:
        type: integer
      dwca_id:
        type: string
      sourcefileid:
        type: string
      occurrenceid:
        type: string
      canonical:
        type: string
      taxon_rank:
        type: string
      country_name:
        type: string
      county_name:
        type: string
      locality:
        type: string
      year_start:
        type: integer
        minimum: 0000
        maximum: 9999
      month_start:
        type: integer
        minimum: 01
        maximum: 12
      day_start:
        type: integer
        minimum: 00
        maximum: 31
      year_end:
        type: integer
        minimum: 0000
        maximum: 9999
      month_end:
        type: integer
        minimum: 01
        maximum: 12
      day_end:
        type: integer
        minimum: 00
        maximum: 31
      eventdate_start:
        type: string
        pattern: '^(\d{4}-\d{2}-\d{2})$'
      eventdate_end:
        type: string
        pattern: '^(\d{4}-\d{2}-\d{2})$'
      verbatim_elevation:
        type: string
      minimum_elevation:
        type: integer
      maximum_elevation:
        type: integer
      department_name:
        type: string
      recorded_by:
        type: string
      habitat:
        type: string
      location:
        $ref: '#/definitions/location'
      taxonomy:
        $ref: '#/definitions/taxonomy'
      provider:
        $ref: '#/definitions/provider'
      resource:
        $ref: '#/definitions/resource'
      institution:
        $ref: '#/definitions/institution'
      collection:
        $ref: '#/definitions/collection'
      catalog:
        $ref: '#/definitions/catalog'
      basis_of_record:
        $ref: '#/definitions/basis_of_record'
  location:
    type: object
    required:
      - lat
      - lon
    properties:
      lat:
        type: number
        format: double
      lon:
        type: number
        format: double
  Facet:
    type: object
    required:
      - field
      - counts
    properties:
      field:
        type: string
      counts:
        type: array
        items:
          type: object
          required:
            - name
            - count
          properties:
            name:
              type: string
            count:
              type: integer
  taxonomy:
    type: object
    properties:
      kingdom_name:
        type: string
      phylum_name:
        type: string
      class_name:
        type: string
      order_name:
        type: string
      family_name:
        type: string
      genus_name:
        type: string
      species_name:
        type: string
      specific_epithet:
        type: string
      infraspecific_epithet:
        type: string
  provider:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      address:
        type: string
      city:
        type: string
      website_url:
        type: string
      logo_url:
        type: string
      email:
        type: string
      phone:
        type: string
  resource:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      title:
        type: string
      description:
        type: string
      intellectual_rights:
        type: string
      citation:
        type: string
      logo_url:
        type: string
      publication_date:
        type: string
        pattern: '^(\d{4}-\d{2}-\d{2})$'
      gbif_package_id:
        type: string
      alternate_identifier:
        type: string
      language:
        type: string
      keyword:
        type: string
      keyword_thesaurus:
        type: string
      hierarchy_level:
        type: string
  institution:
    type: object
    properties:
      code:
        type: string
  collection:
    type: object
    properties:
      id:
        type: string
      code:
        type: string
      name:
        type: string
  catalog:
    type: object
    properties:
      number:
        type: string
  basis_of_record:
    type: object
    properties:
      name:
        type: string
  county:
    required:
      - country_code
      - department_name
      - country_name
      - iso_county_code
      - county_name
      - iso_department_code
    properties:
      country_code:
        type: string
      department_name:
        type: string
      country_name:
        type: string
      iso_county_code:
        type: string
      county_name:
        type: string
      iso_department_code:
        type: string
  department:
    required:
      - country_code
      - lng
      - department_name
      - country_name
      - min_latitude
      - max_longitude
      - iso_department_code
      - lat
      - max_latitude
      - min_longitude
    properties:
      country_code:
        type: string
      lng:
        type: string
      department_name:
        type: string
      country_name:
        type: string
      min_latitude:
        type: number
        format: double
      max_longitude:
        type: number
        format: double
      iso_department_code:
        type: string
      lat:
        type: string
      max_latitude:
        type: number
        format: double
      min_longitude:
        type: number
        format: double
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
      description:
        type: string